added capitalize BPM, capitalize words, and replace underscores features. created about and source link.

+++ src/candlecleaner.py
@@ -14,11 +14,13 @@
 
 import os
 import re
+import webbrowser
 try:
     import tkinter as tk
     from tkinter import *
     from tkinter import filedialog, messagebox, ttk
-    from PIL import Image, ImageTk
+    from PIL.Image import open as popen
+    from PIL.ImageTk import PhotoImage
 except ImportError:
     print("Tkinter is not installed. Please install it before running this script.")
     exit()
@@ -43,49 +45,39 @@ class CleanerApp(tk.Tk):
             icon_path = os.path.join(script_dir, "icon.ico")
 
             # Load the image file using PIL
-            image = Image.open(icon_path)
-            photo = ImageTk.PhotoImage(image)
+            image = popen(icon_path)
+            photo = PhotoImage(image)
 
             # Set the photo as the icon image
             self.iconphoto(True, photo)
 
-        self.menu = tk.Menu(self)
-        self.config(menu=self.menu)
-
-        self.file_menu = tk.Menu(self.menu, tearoff=0)
-        self.menu.add_cascade(label="File", menu=self.file_menu)
-
-        self.file_menu.add_command(label="Select Directory", command=self.select_directory)
-        self.file_menu.add_separator()
-        self.file_menu.add_command(label="Exit", command=self.quit)
-
-        self.options_menu = tk.Menu(self.menu, tearoff=0)
-        self.menu.add_cascade(label="Options", menu=self.options_menu)
-
-        self.scroll_sync_var = tk.BooleanVar(value=True)
-        self.options_menu.add_checkbutton(label="Scroll Sync", variable=self.scroll_sync_var, command=self.scroll_sync_toggle, state='active')
-
-        self.smart_update_var = tk.BooleanVar()
-        self.options_menu.add_checkbutton(label="Candle Clean", variable=self.smart_update_var, command=self.update_file_list)
-
-        self.directory_frame = tk.Frame(self)
-        self.directory_frame.grid(row=1, column=1, pady=10, padx=10, sticky="we")
-        self.string_frame = tk.Frame(self)
-        self.string_frame.grid(row=1, column=2, pady=10, padx=10, sticky="we")
-
         self.directory_var = tk.StringVar()
         self.string_var = tk.StringVar()
         self.replace_var = tk.StringVar()
+        self.scroll_sync_var = tk.BooleanVar(value=True)
+        self.leading_zero_var = tk.BooleanVar(value=True)
+        self.smart_update_var = tk.BooleanVar()
+        self.upper_bpm_var = tk.BooleanVar()
+        self.capitalize_var = tk.BooleanVar()
+        self.underscore_var = tk.BooleanVar()
         self.regex_list = []
 
-        self.directory_label = tk.Entry(self.directory_frame, textvariable=self.directory_var, bd=5)
-        self.directory_label.pack(side="right", fill=X, expand=True)
+        self.directory_frame = tk.Frame(self)
+        self.directory_frame.grid(row=1, column=1, pady=10, padx=10, sticky="we")
+
+        self.string_frame = tk.Frame(self)
+        self.string_frame.grid(row=1, column=2, pady=10, padx=10, sticky="we")
 
         self.to_replace = tk.Frame(self.string_frame, bg=self.directory_frame["bg"])
         self.to_replace.pack(side="left")
 
         self.replacer = tk.Frame(self.string_frame, bg=self.directory_frame["bg"])
         self.replacer.pack(side="left")
+
+        self.spacer = tk.Label(self.directory_frame, text="", padx=5)
+        self.spacer.pack(side="right")
+        self.directory_label = tk.Entry(self.directory_frame, textvariable=self.directory_var, bd=5)
+        self.directory_label.pack(side="right", fill=X, expand=True)
 
         self.string_label = tk.Label(self.to_replace, text="Text to replace:", padx=5)
         self.string_label.pack(side="left", anchor="center")
@@ -101,7 +93,7 @@ class CleanerApp(tk.Tk):
         self.replace_entry.pack(side="left", anchor="w")
         self.replace_entry.config(state='normal')
 
-        self.rename_button = tk.Button(self, text="Rename Files", command=self.rename_files, justify=CENTER)
+        self.rename_button = tk.Button(self, text="Rename Files", command=self.rename_files, justify=CENTER, bd=3)
         self.rename_button.grid(row=4, column=2, pady=10)
 
         # Associate the validation function with the string variable
@@ -123,15 +115,42 @@ class CleanerApp(tk.Tk):
         self.updated_file_tree.column("size", width=100)
         self.updated_file_tree.grid(row=3, column=2, padx=10, pady=10, sticky="nsew")
 
-        self.file_tree_scroll = tk.Scrollbar(self, orient="vertical", command=self.file_tree_yview)
+        self.file_tree_scroll = tk.Scrollbar(self, orient="vertical")
         self.file_tree_scroll.grid(row=3, column=1, rowspan=1, padx=10, pady=10, sticky="nse")
         self.file_tree.configure(yscrollcommand=self.file_tree_scroll.set)
         self.file_tree.bind("<MouseWheel>", self.file_tree_scroll_mouse_wheel)
 
-        self.updated_file_tree_scroll = tk.Scrollbar(self, orient="vertical", command=self.updated_file_tree_yview)
+        self.updated_file_tree_scroll = tk.Scrollbar(self, orient="vertical")
         self.updated_file_tree_scroll.grid(row=3, column=2, rowspan=1, padx=10, pady=10, sticky="nse")
         self.updated_file_tree.configure(yscrollcommand=self.updated_file_tree_scroll.set)
         self.updated_file_tree.bind("<MouseWheel>", self.updated_file_tree_scroll_mouse_wheel)
+
+        self.menu = tk.Menu(self)
+        self.config(menu=self.menu)
+
+        self.file_menu = tk.Menu(self.menu, tearoff=0)
+        self.menu.add_cascade(label="File", menu=self.file_menu)
+        self.file_menu.add_command(label="Select Directory", command=self.select_directory)
+        self.file_menu.add_separator()
+        self.file_menu.add_command(label="Exit", command=self.quit)
+
+        self.options_menu = tk.Menu(self.menu, tearoff=0)
+        self.menu.add_cascade(label="Options", menu=self.options_menu)
+        self.options_menu.add_checkbutton(label="Scroll Sync", variable=self.scroll_sync_var, command=self.scroll_sync_toggle, state='active')
+
+        self.cleaner_menu = tk.Menu(self.menu, tearoff=0)
+        self.menu.add_cascade(label="Candle Cleaner", menu=self.cleaner_menu)
+        self.cleaner_menu.add_checkbutton(label="Enable", variable=self.smart_update_var, command=lambda : self.update_file_list(True))
+        self.cleaner_menu.add_separator()
+        self.cleaner_menu.add_checkbutton(label="Including Leading 0s", variable=self.leading_zero_var, state='disabled', command=lambda : self.update_file_list(True))
+        self.cleaner_menu.add_checkbutton(label="Capitalize BPM", variable=self.upper_bpm_var, state='disabled', command=lambda : self.update_file_list(True))
+        self.cleaner_menu.add_checkbutton(label="Capitalize Words", variable=self.capitalize_var, state='disabled', command=lambda : self.update_file_list(True))
+        self.cleaner_menu.add_checkbutton(label="Replace Underscores", variable=self.underscore_var, state='disabled', command=lambda : self.update_file_list(True))
+
+        self.help_menu = tk.Menu(self.menu, tearoff=0)
+        self.menu.add_cascade(label="Help", menu=self.help_menu)
+        self.help_menu.add_command(label="About", command=self.show_help)
+        self.help_menu.add_command(label="Source", command=self.show_source)
 
         # Add a callback to each scrollbar that sets the position of the other scrollbar
         def sync_scrolls(first_treeview, second_treeview):
@@ -160,20 +179,20 @@ class CleanerApp(tk.Tk):
             self.file_tree.unbind("<MouseWheel>")
             self.updated_file_tree.unbind("<MouseWheel>")
 
-    def file_tree_yview(self, *args):
-        self.file_tree.yview(*args)
-        self.updated_file_tree.yview(*args)
-
-    def updated_file_tree_yview(self, *args):
-        self.updated_file_tree.yview(*args)
-        self.file_tree.yview(*args)
-
     # If checkbox is selected, string entry should be disabled
     def validate_string_entry(self):
         if self.smart_update_var.get():
+            self.cleaner_menu.entryconfig(2, state='normal')
+            self.cleaner_menu.entryconfig(3, state='normal')
+            self.cleaner_menu.entryconfig(4, state='normal')
+            self.cleaner_menu.entryconfig(5, state='normal')
             self.string_entry.config(state='disabled')
             self.replace_entry.config(state='disabled')
         else:
+            self.cleaner_menu.entryconfig(2, state='disabled')
+            self.cleaner_menu.entryconfig(3, state='disabled')
+            self.cleaner_menu.entryconfig(4, state='disabled')
+            self.cleaner_menu.entryconfig(5, state='disabled')
             self.string_entry.config(state='normal')
             self.replace_entry.config(state='normal')
 
@@ -190,7 +209,7 @@ class CleanerApp(tk.Tk):
     def update_file_list(self, make_regex = False):
         directory_path = self.directory_var.get()
         smart_update = self.smart_update_var.get()
-        #self.regex_list = []
+        if make_regex: self.regex_list = []
         self.validate_string_entry()
         if os.path.isdir(directory_path):
             # Clear the file tree
@@ -228,7 +247,7 @@ class CleanerApp(tk.Tk):
                         parent_ids[sub_dirpath] = {'file_tree': sub_parent_id, 'updated_file_tree': updated_sub_parent_id}
 
                 # Create regex list for selected directory
-                if make_regex:                       
+                if make_regex:
                     if filenames:
                         directory_name = os.path.basename(dirpath)
                         self.generate_regex(filenames, directory_name)
@@ -253,6 +272,9 @@ class CleanerApp(tk.Tk):
                             pattern = re.compile(re.escape(string_to_remove), re.IGNORECASE)
                             updated_file_name = re.sub(pattern, replacement, file_shown)
                             updated_file_size = os.path.getsize(file_path)
+                            if self.upper_bpm_var.get(): updated_file_name = updated_file_name.replace('bpm', 'BPM')
+                            if self.capitalize_var.get(): updated_file_name = self.capitalize_string(updated_file_name)
+                            if self.underscore_var.get(): updated_file_name = self.replace_underscore(updated_file_name)
                             self.updated_file_tree.insert(parent_ids[dirpath]['updated_file_tree'], END, text=updated_file_name, values=(updated_file_size,))
                         else:
                             self.updated_file_tree.insert(parent_ids[dirpath]['updated_file_tree'], END, text=filename, values=(file_size,))
@@ -262,35 +284,44 @@ class CleanerApp(tk.Tk):
         smart_update = self.smart_update_var.get()
         success = True
 
-        if os.path.isdir(directory_path):
-            for dirpath, dirnames, filenames in os.walk(directory_path):
-                for filename in filenames:
-                    if filename != ".DS_Store" and filename != "":
-                        if smart_update:
-                            current_directory = os.path.basename(dirpath)
-                            current_regex = next((regex[0] for regex in self.regex_list if regex[1] == current_directory), "")
-                            string_to_remove = current_regex
-                            file_shown = re.sub(r'[ \-_]+', '_', os.path.splitext(filename)[0].lower()) + os.path.splitext(filename)[1]
-                            replacement = ""
-                        else:
-                            string_to_remove = self.string_var.get()
-                            replacement = self.replace_var.get()
-                            file_shown = filename
-                        pattern = re.compile(re.escape(string_to_remove), re.IGNORECASE)
-                        updated_file_name = re.sub(pattern, replacement, file_shown)
-                        src = os.path.join(dirpath, filename)
-                        dst = os.path.join(dirpath, updated_file_name)
-                        try:
-                            os.rename(src, dst)
-                        except:
-                            success = False
-                            messagebox.showerror("Failed" "Couldn't rename file: " + file_shown)
+        if messagebox.askyesno("Confirmation", "Rename Files?"):
+            if os.path.isdir(directory_path):
+                for dirpath, dirnames, filenames in os.walk(directory_path):
+                    for filename in filenames:
+                        if filename != ".DS_Store" and filename != "":
+                            if smart_update:
+                                current_directory = os.path.basename(dirpath)
+                                current_regex = next((regex[0] for regex in self.regex_list if regex[1] == current_directory), "")
+                                string_to_remove = current_regex
+                                file_shown = re.sub(r'[ \-_]+', '_', os.path.splitext(filename)[0].lower()) + os.path.splitext(filename)[1]
+                                replacement = ""
+                            else:
+                                string_to_remove = self.string_var.get()
+                                replacement = self.replace_var.get()
+                                file_shown = filename
+                            pattern = re.compile(re.escape(string_to_remove), re.IGNORECASE)
+                            updated_file_name = re.sub(pattern, replacement, file_shown)
+                            if self.upper_bpm_var.get(): updated_file_name = updated_file_name.replace('bpm', 'BPM')
+                            if self.capitalize_var.get(): updated_file_name = self.capitalize_string(updated_file_name)
+                            if self.underscore_var.get(): updated_file_name = self.replace_underscore(updated_file_name)
+                            src = os.path.join(dirpath, filename)
+                            dst = os.path.join(dirpath, updated_file_name)
+                            try:
+                                os.rename(src, dst)
+                            except:
+                                success = False
+                                messagebox.showerror("Failed" "Couldn't rename file: " + file_shown)
+                self.smart_update_var.set(False)
+                self.update_file_list()
+                if success:
+                    messagebox.showinfo("Success", "Files renamed successfully!")
 
-            self.update_file_list()
-            if success:
-                messagebox.showinfo("Success", "Files renamed successfully!")
+    def show_help(self):
+        messagebox.showinfo("About", "Mass file renamer, specifically geared toward audio sample libraries.")
+
+    def show_source(self):
+        webbrowser.open_new_tab("https://github.com/SuluCandles/candlecleaner")
             
-
     def generate_regex(self, filenames, dirname):
             # Filter out any filenames that are not hidden
             filenames = [f for f in filenames if not f.startswith(".")]
@@ -302,17 +333,37 @@ class CleanerApp(tk.Tk):
             # and removing the file extension from each filename
             normalized_filenames = [re.sub(r'[ \-_]+', '_', os.path.splitext(f)[0].lower()) for f in filenames]
             common_prefix = os.path.commonprefix(normalized_filenames)
-            common_prefix = common_prefix.removesuffix('0')
+            if not common_prefix:
+                common_prefix = "*"
+            else:
+                if self.leading_zero_var.get():
+                    common_prefix = common_prefix.removesuffix('0')
             prefix_regex = re.escape(common_prefix).replace('\\_', '[ _-]')
 
-            pattern = prefix_regex# + ".*"
-            regex = pattern.replace('\\-', '-')
+            regex = prefix_regex.replace('\\-', '-')
             # Add the regex and the directory to the regex list
             self.regex_list.append((regex, dirname))
+
+    def capitalize_string(self, filename):
+        words = filename.split("_")
+        capitalized_words = []
+        if words[0].lower() == "the":
+            words[0] = words[0].capitalize()
+        for word in words:
+            if not (word.lower() in {"and", "the", "of", "or", "a", "an", "in", "to", "for", "with", "on", "at", "by", "but", "nor", "from", "bpm"}):
+                word = word.capitalize()
+            capitalized_words.append(word)
+        output_string = "_".join(capitalized_words)
+        return output_string
+
+    def replace_underscore(self, filename):
+        words = filename.split("_")
+        words = " ".join(words)
+        return words
 
 if __name__ == '__main__':
     # create the application instance
     app = CleanerApp()
     
     # start the main event loop
-    app.mainloop()            
+    app.mainloop()
